Rename Postgres to PostgreSQL, move default implementations to separate files
